[FIX] UserDetailsImpl fix

See merge request s03-webmobile1-sub3/s03p13a502!23

+++ Plog_Backend/src/main/java/com/ssafy/plog/service/UserDetailsImpl.java
@@ -15,7 +15,7 @@ import com.ssafy.plog.models.User;
 public class UserDetailsImpl implements UserDetails {
 	private static final long serialVersionUID = 1L;
 
-	private Long id;
+	private int id;
 
 	private String username;
 
@@ -30,7 +30,7 @@ public class UserDetailsImpl implements UserDetails {
 	
 	private Collection<? extends GrantedAuthority> authorities;
 
-	public UserDetailsImpl(Long id, String username, String email, String password, String birthday, String phone,
+	public UserDetailsImpl(int id, String username, String email, String password, String birthday, String phone,
 			Collection<? extends GrantedAuthority> authorities) {
 		this.id = id;
 		this.username = username;
@@ -61,7 +61,7 @@ public class UserDetailsImpl implements UserDetails {
 		return authorities;
 	}
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
